+++ src/collective/liveblog/content.py
@@ -8,6 +8,11 @@ class Liveblog(Container):
 
     """A liveblog is a blog post which is intended to provide a rolling
     textual coverage of an ongoing event.
+
+    The _last_microupdate_deletion attribute is used to detect if a hard
+    refresh of the views is needed.
     """
 
     grok.implements(ILiveblog)
+
+    _last_microupdate_deletion = '0.0'